+++ packages/rest-explorer/src/rest-explorer.controller.ts
@@ -33,12 +33,17 @@ export class ExplorerController {
   }
 
   indexRedirect() {
-    this.response.redirect(301, this.request.url + '/');
+    const url = this.request.originalUrl || this.request.url;
+    this.response.redirect(301, url + '/');
   }
 
   index() {
+    let openApiSpecUrl = this.openApiSpecUrl;
+    if (this.request.baseUrl && this.request.baseUrl !== '/') {
+      openApiSpecUrl = this.request.baseUrl + openApiSpecUrl;
+    }
     const data = {
-      openApiSpecUrl: this.openApiSpecUrl,
+      openApiSpecUrl,
     };
 
     const homePage = templateFn(data);