+++ packages/boot/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.13...@loopback/boot@1.0.14) (2019-03-01)
+
+**Note:** Version bump only for package @loopback/boot
+
+
+
+
+
+## [1.0.13](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.12...@loopback/boot@1.0.13) (2019-02-25)
+
+**Note:** Version bump only for package @loopback/boot
+
+
+
+
+
 ## [1.0.12](https://github.com/strongloop/loopback-next/compare/@loopback/boot@1.0.11...@loopback/boot@1.0.12) (2019-02-08)
 
 **Note:** Version bump only for package @loopback/boot